libpmi: drop PMIX heuristic
Problem: The logic that dlopens libpmix.so, expecting to
find PMI-1 symbols there if PMIX_SERVER_URI is set in
the environment no longer works since PMI-1 symbols
were moved to a separate compatibility library.

On the LSF/jsrun based IBM system where PMIX is the native
PMI, we have requested that IBM supply the libpmi.so compatability
library.  Once they do, flux should be able to find it in the
usual way.

Meanwhile delete the PMIX heuristic.

Fixes #918