build: substitute runstatedir in autoconf < 2.70
Problem: flux clients cannot connect to the system
instance using compiled-in path.

The compiled in path is prefixed with $runstatedir,
but this variable is empty in versions of autoconf
before 2.70.

Add a conditional to configure.ac so that runstatedir
is substituted to /run in Makefile.am's if $runstatedir
is not set.

Fixes #1573