wreck: allow -o mpi=TYPE option for wreckrun and submit
Problem: MPI specific variables and other setup is currently implemented
via Lua plugins for maximum flexibility, but these plugins are always
active, which may cause unnecessary conflicts or other environmental cruft.

Add a new `-o mpi=TYPE` option to wreckrun and submit, which can be used
to selectively opt-in to certain MPI types that have particularly onerous
setup requirements.

The variable can be checked with `wreck:getopt ("mpi")` within the Lua
plugin specific to a given MPI type, and the setup aborted if the given
MPI type is not selected.