libpmi: add pmix_client implementation
Extend Flux PMI client to allow PMIx to be used.
Support for this mode is only built when flux-core
is configured --with-pmix.

The purpose is to allow Flux to boot on systems
that provide libpmix.so but do not ship the pmix
PMI-1 compatibility libraries, such as IBM system
based on LSF/jsrun.

This code is based on the pmix project's PMI-1
compatibility library code, modified to fit Flux's
internal PMI library structure.

The new conditional dependency on pmix necessitates
the following build system changes:

- Link libpmi.la and libpmi2.la against libpmix.so.
- Drop libpmi.la from libflux-internal.la library.
- Link pmi servers (wrexecd, flux-start, tests)
  against src/common/libpmi/libpmi.la
- Link pmi clients (broker, tests)
  against src/common/libpmi.la.

Fixes #1581