libkvs/getroot: add new API functions for getroot
Add API functions for sending a kvs.getroot RPC
to the kvs service to get the current root of a
KVS namespace.

If the FLUX_KVS_WATCH flag is set, divert the RPC
to the kvs-watch service, which in addition to
responding with the current root, will respond
each time the root changes after that.

flux_kvs_getroot_cancel() sends a cancellation
request to "unwatch" a namespace.  Unlike
flux_kvs_unwatch(), which finds the request to
cancel by sender and KVS key, this function
cancels by sender and matchtag, which can be
more precise.