wreck: add tests for several lua mpi plugins
Specifically tests the intel and spectrum mpi "personalities".

For intel mpi, test that when the I_MPI_PMI_LIBRARY environment
variable is set, it is overwritten with the path to Flux's library.

For spectrum mpi, test that the MPI_ROOT environment variable is set
to the expected path and that the soft stack limit is set to 10240.