wreck/lua: fix crash due to Lua stack overflow
Problem: The Lua stack is not reset anywhere in luastack.c when
loading and running Lua wrexecd plugins. We got away with this
when each script was run in its own lua_thread (coroutine), because
the stacks were not shared between scripts, but when the use of
coroutines was removed, the stack growth was large enough to cause
crashes in some tests that ran many tasks.

Fix the bug by resetting the Lua stack before executing a callback
function in each lua script.

Fixes #1592