t/Makefile.am: fix make check failure when LUA_PATH not set
Problem: In t/Makefile.am, `./?.lua` is prepended to LUA_PATH in
order to ensure that the build tree's `fluxomter.lua` is found,
even if LUA_PATH doesn't contain the current working directly.
This doesn't work if LUA_PATH is currently empty, since you end
up with *only* `./?.lua` in LUA_PATH.

Since the default LUA_PATH already contains the current dir, we
only need to prepend to it if it is already set. So, do nothing
when LUA_PATH not set.