t2000-wreck: fix prereq test for MULTICORE
Problem: The test for MULTICORE prepreq in the wreck tests was
accidentally broken in f9f4baea, such that the cpu-affinity tests
were no longer running on most systems.

Fix the sense of the test for MULTICORE support so that these tests
are run, and fix a couple bugs in the tests detected now that the
tests actually are executed.