wreck: simplify fake R_lite generation in wreckrun
Problem: Over time the code to support "fake" resource allocations
in flux-wreckrun (used when a scheduler is not present or with
-I, --immediate) has grown complex and difficult to follow/maintain.

Now that R_lite is the only allocation method supported by wreck,
simplify generation of fake R_lite in flux-wreckrun. As a side-effect
support the --cores-per-task option, which was previously ignored.