modules/kvs-watch: handle key lookups
Implement kvs-watch.lookup.

The 'struct watcher' is extended with fields to support
the key and state associated its lookup.

At each root update, the watched key is compared
against the key array transmitted in the kvs.setroot
event.  If a match is found, initiate a kvs.lookup
on the key, and direct the result to the user in a
response message.

The futures associated with the kvs.lookup requests are
stored in a zlist so that commit order can be preserved
even if lookup responses arrive out of order.

There is room for future optimization in this code:
1) Each root commit triggers the traversal of the
list of watchers.  This could be sped up with a hash
of keys, per namespace.
2) Multiple watchers of the same key are independent,
and will trigger a seperate kvs.lookup request per
watcher, per commit.

N.B. because a kvs.lookup can cross into a new namespace,
kvs.lookup RPCs are made with the watching user's creds.
This must be considered when optimizing 2) above.
Perhaps in the future, namespace + owner could be
returned in lookup responses.