connector/local: fix truncated output warning from GCC 8
Problem: compilation of connector/local.c was failing under GCC 8
due to a warning (fatal, due to -Werror) about output truncation when
the variable sockpath (an array of size PATH_MAX + 1) is copied to
addr.sun_path (107 bytes).

Since Unix domain socket pathnames are actually constrained to the length
of  struct sockaddr_un.sun_path array, use this length instead of PATH_MAX
in sizing the array in which the socket path is constructed, so that
truncation error is caught in connector_init() instead of actually causing
truncation when copying in connect_sock_with_retry().

Fixes #1633