tests: fixes for GCC 8 truncation warnings
Problem: many tests fail to compile on GCC 8 due to a truncation
warning (fatal due to use of -Werror), when using snprintf() to
construct a string where a source buffer of the string is the same
size or larger than the destination string.

Technically, this isn't such a big deal, and because of the tests
the truncation would actually never occur, but it is easiest for
now to just resize some of the source buffers to be smaller than the
destination and suppress the warnings.