libflux/rpc: add flux_rpc_message()
Problem: there is no mechanism to set the userid and rolemask
in a message sent with the flux_rpc_*() functions.

The ability to set the rolemask and userid of a message is
useful when sending a request on behalf of another user.
It only has an effect when running as the instance owner,
since otherwise, the connector rewrites the credentials
with those of the authenticated user.

To make this possible, expose low level RPC constructor
that accepts a request message as a parameter.  It is
sent out verbatim except for setting the matchtag (per
rpc flags) and nodeid.