add support for asciidoctor as manpage generator
This allows asciidoctor to be used in place of a2x, which removes the
dependency on goodness-knows what docbook package our a2x generator
depends on that isn't in either our dependency list or theirs.  This
exists primarily because after making three versions of a docker image
trying to get one that could build the docs the other way, this works
faster, with less deps, and seemingly the same result.