libkvs/eventlog: add RFC 18 KVS event log support
Add functions for manipulating RFC 18 KVS event logs,
which consist of newline-terminated events that look like:

timestamp name [context ...]\n

Where timestamp is seconds since UNIX Epoch UTC, with
subsecond precision, name is a single word event name,
and context is free form text that may include arguments
or human readable log data, depending on how the eventlog
is being used.

The eventlog API is decoupled from KVS lookup and txn API's
so that it can be indepedently tested.  Higher level
API functions may prove useful later on, but for now let's
start with the basics.

The eventlogs support multiple un-coordinated writers,
utilizing the FLUX_KVS_APPEND operation to update the log
without rewriting it.  Appends are compatible with the
new FLUX_KVS_WATCH flag as well, since each append is
a direct operation on a key that is attached to the KVS
setroot event message.

The entire eventlog snapshot is returned by a KVS lookup.
There is currently no KVS mechanism to receive only the
updates to the log.  Therefore, a flux_kvs_eventlog_update()
function is provided which appends only the new log entries
to a struct flux_kvs_eventlog object.  Pointers to events
obtained through the iterators before an update remain valid
after the update.

To avoid the need to allocate memory, the event encode/decode
functions copy the name and optional context to fixed length
buffers.  The maximum length of these fields was not initially
specified in RFC 18, so that will need to be addressed if
the fixed limits are acceptable.