test: add script to run travis checks in docker
Add a helper script for creating docker images for testing,
mainly meant to be used with travis (but will work outside of
travis)

This simplifies the travis use case, and allows testing of
docker image builds from the command line during development.

This change updates the "travis_run.sh" script to take configure
arguments on the command line, as well as other changes and
improvements.