t/lua: ensure non-installed lalarm module can be found
Problem: With TEST_INSTALL=t tests, the installed fluxometer
cannot find the non-installed lalarm.so Lua module, used by
some of the tests.

Make sure that the build directory LUA_CPATH is in the LUA_CPATH
used by `make check`, so that lalarm is found at runtime during
use of the "installed" flux-core.