cmd/flux: do not die on EACCES to build dir
Problem: if the original build directory for flux-core is still
around and inaccessible to the flux(1) command due to permissions,
the flux executable dies with

 flux realpath (private/path) Permission denied

while attempting to determine if it is run out of a builddir or the
installed path.

Fix this by checking for either ENOENT or EACCES from realpath(3),
and falling through to the installed path check if either error
is encountered.

Fixes #1674