build: fix --disable-docs logic error
Problem: configure --disable-docs no longer works.

The recent change to prefer a2x over asciidoctor broke
logic for --disable-docs.  If docs are not enabled,
ADOC is unset, but the logic requires it to be
set to "not-found" to disable docs building.

Drop the default "not-found" value in AC_CHECK_PROGS(),
and check for ADOC unset.

Tested
- with a2x installed, docs enabled
- with a2x installed, docs disabled
- with nothing installed, docs enabled

Fixes #1677