travis-ci: fix docker image deploy
Problem: Docker image auto-deployment for new merges to master or
tags was broken due to a testing "repo slug" left in the test
to ensure we were running in the official flux-core repository.

Update the repo slug to `flux-framework/flux-core`.

Also, include the DOCKER_TAG=t variable for CentOS 7 builds so
we get a centos docker image.

Finally, update the names of builders in Travis-CI for clarity.