t2201-job-cmd.t: fix test for HAVE_FLUX_SECURITY
Problem: t2201-job-cmd.t sets a HAVE_FLUX_SECURITY prereq by
checking for the `sign-type` arg on the flux-job command.
However, this arg exists on `flux job submitbench` not
just flux-job, so the test was failing.

Adjust the test to use `flux job submitbench` so the security
tests are not skipped.