libjob: add flux_job_list()
Add API function for retrieving a list of active jobs
from the job manager.

The request includes a maximum number of jobs to receive
and a list of attributes.

The response is an array of job objects, where each object
is a dictionary of the requested attributes.

Valid attributes thus far are:
id - the jobid (64 bit)
userid - the submitting userid (32 bit)
priority - submitting priority (int [0:31])