docker: support not mounting the host home directory
Adds a `--no-home` option to the `docker-run-checks.sh` script.

Useful for building on Mac OSX where mounting the home directory within a Linux
container can cause issues down the road.  For example, the $HOME in OSX is
`/Users/$username`, but in linux, it is `/home/$username`.  Some tools, like
ssh-ident and emacs, can create files in your home directory with absolute path
information in them.  Launching these tools within the linux docker container
while the host Mac OSX home directory is mounted can lead to inconsistent state.

Additionally, if you are a hyper-conservative user, this limits the exposure and
risk of running the script.