tap-driver.sh: add quiet mode
Problem: for CI testing and other environments, one line of TAP
output per test is becoming a bit verbose. This is causing an issue
in the Travis-CI environment specifically, where builds are failing
with truncated output, possibly due to large volume of test output.

Add a "quiet" mode to tap-driver.sh that suppresses PASS lines of
TAP output and issues a summary at the end of each test. This mode
can be initiated by setting TAP_DRIVER_QUIET=1 in the environment
before `make check`.