pylint: enable pylint checking in travis
Disable unwanted and address relevant lint warnings.

Bad first argument given to super class:

The biggest change was switching `super(self.__class__, self)` to
actually directly reference the class.  `self.__class__` works when the
class is at the front of the MRO, but as soon as another class inherits
from that class, an infinite recursion happens in the `__init__`.  Most
of the uses were in a class that was inside another class, so I doubt
anyone would actually subclass them, but at least now pylint is happy.

no-else-return:

For this error, in most cases, I obliged the linter, but in one
particular case (src/bindings/python/flux/wrapper.py:224), I think the
if/else is sufficiently large that the way it's currently written is
easier to read/parse.  So I overrode that case.

no-value-for-parameter:

pylint really hates the `ffi` and `lib` modules being passed in as
arguments to `Wrapper.__init__`.  I don't see a better way to do it, so
I just manually overrode each instance.

useless-object-inheritance:

pylint3 doesn't like `class WrapperBase(object)`.  Disabled universally
since this is required for py2 backwards compatibility.