libjob: add FLUX_JOB_PRE_SIGNED flag
Problem: users of flux_job_submit() must know whether
flux-core was built --with-flux-security or not.

Add a FLUX_JOB_PRE_SIGNED flag that allows jobspec to
be signed before submission, but change the default
so flux_job_submit() signs the jobspec internally.

If built --with-flux-security, configure a security context using
default config file path and cache it in the flux_t handle.
Each time flux_job_submit() is called, the context is
reused, and the jobspec is signed with the default mechanism.

Take special care to enable the user to access a textual
error message from flux-security upon failure.  If a textual
error message is available, instead of returning NULL,
arrange for flux_job_submit() to return a future pre-fulfilled
with the error string.

If built without --with-flux-security, sign unconditionally
with the simplified local "sign_none" class.

Update flux-job to use the FLUX_JOB_PRE_SIGNED flag if
built --with-flux-security and the user requests non-default
security options that can only be set by signing the jobspec
with full access to the flux-security API.

Fixes #1715