modules/job-ingest: unlink KVS entry on submit error
Problem: job entries remain in the KVS in three cases
where job-ingest.submit fails.

Unlink the KVS entries associated with a batch of submit
requests if the job-manager.submit RPC fails (two cases)
or if we fail to set up the continuation for the KVS commit.

Perform the unlink in parallel with responding to the user.
This makes writing a test specifically for this case difficult
in the standalone ingest sharness test.  However, a future
job-manager test that loads the job manager after some ingest
ENOSYS failures should find unexpected jobs in its queue if
the KVS entries remain afterwards.

Rather than make the cleanup code more complicated so it
completes before responding to the user, keep it simple and
defer testing to the job manager.