cmd/flux-job: avoid unnecessary jobspec copy
Problem: when not pre-signing jobspec, flux job submitbench
copies the jobspec read by read_all() in order to add a
terminating NULL for flux_job_submit().  Recently read_all()
was modified to add a terminating NULL that is not reflected
in the size, so this extra copy is no longer needed.

Remove the extra copy.