broker: drop munge encryption on event overlay
Problem: flux-core's munge dependency is inconvenient
for spack users building single-user flux, because
the socket path in munge client library and server
can easily be mismatched.

Munge is only needed to provide privacy for EPGM
event distribution, which is not generally used since
event distribution defaults to the TBON.  Munge was probably
not the best choice anyway, given the RTT to the munge
daemon for every message - encrypting with libsodium would
be better.

Disable encryption of PUB/SUB messages for the time being
to resolve the inconvenience.  Revisit when EPGM is deployed
for the system instance.