modules/pymod: register service name dynamically
Add code to pymod to register a service name dynamically based on the
name of the script being loaded. This allows a python-based module
to register any service name, not just `pymod` and handle request
messages for that service.

For example, an echo server may be loaded with

  flux module load pymod echo

and process messages matching topic string "echo.*".

Code to *deregister* the automatic `pymod` service name is also
included in this commit but commented out, because, though now
theoretically possible, it is not currently safe to load a second
pymod module with a different python script (and there is no way to
unload just one of two modules with the same name)