modules/pymod: ensure libpython syms are global
Problem: On some distributions (e.g. Ubuntu 18.04) _cffi_backend.so
is not linked against libpython DSO, and since pymod.so is loaded
with RTLD_LOCAL, symbols from libpython may not be visble to cffi,
and thus `import cffi` fails with unresolved symbol errors. (This
may be a general problem with any python module that is not
linked with libpython in this scenario)

To fix, promote libpython symbols to global by calling dlopen(3)
on libpython (as found by configure) with RTLD_GLOBAL.