connector-local: enable routes on all api requests
Problem: The connector-local module throws an error and drops any
request messages from connected clients which do not have the
FLUX_MSGFLAG_ROUTE flag. Currently the python flux_send() binding
does not set this flag for requests, so python scripts are unable
to use this interface.

It is arguable that the local connector *should* drop requests
without FLUX_MSG_ROUTE set, because this means that the msg being
sent via the connector is not a valid request. However, since we
know all requests will need FLUX_MSGFLAG_ROUTE set, and the connector
is accepting messages on behalf of clients, is there a harm in
ensuring the route flag is set on all requests? This reduces the
burden of "encoding" requests from the API clients to the API
itself, which may or may not be the correct approach.

Until the python bindings can encode a request, though, add code
to the connector-local module to enable routes on all request
messages.