testsuite: avoid relative paths in PYTHONPATH
Problem: python or one of its modules seems to clear relative
paths from PYTHONPATH, and this is preventing the t9001-pymod.t
test from working problem with a failure to import the 'flux' module.

The problem seems to be resolved using absolute paths to top_srcdir
and top_builddir, so just use this as the solution for now.