modules/kvs-watch: Change namespace remove errno
When a namespace is removed, return ENOTSUP instead of ENOENT
to the user.  This is more consistent with the usage of ENOTSUP
in the main kvs to indicate when a namespace does not exist.  In
addition, it allows the user to differentiate between a namespace
no longer existing and a key no longer existing.