libflux: fix composite futures for rpcs
Problem: composite futures, especially chained composites, do not
currently work for RPCs and probably in many other situations.
This is because the internal reactor and possibly cloned flux_t
handle for a chained future may not match those in the "previous"
future. Thus, a user calling flux_future_get() on a "next" or
parent future may be blocked on a flux_reactor_run() using the
wrong handle, which means a message required by a previous future
in the chain will never arrive.

Fix by ensuring *both* reactor and flux_t handle are propagated
to next, previous, or child futures as necessary during creation and
initialization of composite futures.

Fixes #1750.