rc3: conditionally save kvsroot on shutdown
Problem: although the content backing store can be
made to persist after a Flux instance terminates
by setting 'persist-directory' or 'persist-filesystem'
broker attributes, there is currently no mechanism
to access KVS data in it because the KVS root is
not saved.

Add code to rc3 to save the root in a file called
'kvsroot.final' in the persist-directory, if set.

As discussed in #1754, one can start a flux instance
and make the old instance's final KVS snapshot available
in the new instance's KVS with a command like:

flux kvs put --treeobj \
    oldroot-snapshot=$(cat kvsroot.final)