testsuite: add y2j.py for yaml->json conversion
Add y2j.py which reads YAML on stdin, and
writes JSON on stdout. It can be used to preprocess
YAML sample jobspec for input to existing tests,
in preparation for job-ingest requring jobs to
be submitted as signed JSON rather than allowing
either YAML or JSON.

Invalid YAML triggers a python exception and
a nonzero exit.

JSON output is compact, with no whitespace added
for readability.

Depends on python-yaml.