modules/job-ingest: require signed JSON not YAML
Add code to validate that submitted jobspec is
in JSON form.  If YAML is submitted (or any other
invalid JSON), an EINVAL error response is sent back
to the user, with a textual error message from the
JSON parser.

The jobspec is still passed through the C++ jobspec
validator, which accepts both JSON and YAML.