testsuite: use flux exec with -n when stdin not used
Problem: since flux exec inherits stdin from the shell,
and it nees to put stdin into O_NONBLOCK mode to use
the buffer watchers, there is a risk that the shell's
stdin file descriptor may be left in that mode after
flux exec finishes.

As discussed in #1803, in all tests where flux exec is
used without the need for stdin, add the -n option to
minimize risk of this effect.