libflux/reactor: buffer read/write watchers require O_NONBLOCK
Problem: buffer_read/write_watcher_create() sets the O_NONBLOCK
on its file descriptor, but doesn't restore the original setting
when the buffer is destroyed.  When the fd is stdin or stdout and was
inherited from the shell, leaving it in O_NONBLOCK mode can cause
the next programs run by the shell to fail with EAGAIN.

Since the fd is not "owned" by the buffer_read_watcher, but
is required to be in O_NONBLOCK mode, simply check that the
mode is set on entry and leave management of its final state
to the calling program.

If the flag is not set, fail with EINVAL.

Fixes #1803.

Update unit tests.

Co-authored-by: Mark A. Grondona <mark.grondona@gmail.com>