libutil/popen2: fix close-on-exec flag use
Problem: the popen2() call set close-on-exec flag on the parent
side of the socketpair was using F_SETFL and O_CLOEXEC instead of
F_SETFD and FD_CLOEXEC. Thus the intended result was not applied
to the file descriptor and the parent fd was most likely being
leaked into the child process.

Instead, use the new fd_set_cloexec() call, which is simpler, uses
the correct flags, and has a unit test to ensure the correct flag
is applied.