build: ensure -l, -L linker options are not in LDFLAGS
Problem: libtool appears to get befuddled when -L/path/to/foo -lfoo
type options appear in foobar_LDFLAGS, which is reserved for other
linker flags only. In these cases libtool might move a `-L` option
up in the argument vector, causing incorrect path to be first in the
linker path. Worst case, this can cause errors when trying to
side-install a newer version of flux-core when an old version exists
on the system. The relink will fail with unresolved symbols if a
-L/usr/lib or -L/usr/lib64 gets promoted up the argument vector.

While not a guaranteed fix, ensure that libraries and -L, -l link
options apper only in *_LIBADD or *_LDADD automake variables. This
seems to make it less likely that libtool will permute these
arguments, and works around the issue in most cases.