testsuite: relocate more libflux unit tests
Problem: dispatch, handle, log, reactor, reduce,
and rolemask/loop unit tests are located in the
sharness area.

Now that the test server offers a loopback mode, these
tests that were dependent on the loop:// connector can
be moved to libflux/test.

There was a name clash on the 'handle.c' test, but the
one in libflux only had a couple of checks in it, so
those checks were pulled into the larger loopback one.

There was a name clash on the 'reactor.c' test.
The loopback one was renamed to 'reactor_loop.c'.

Renamed rolemask/loop.c to rpc_security.c.

Partially addresses #1823