cmd/flux-start: ignore SIGTTOU
Problem: when the initial program is an interactive shell,
the "current terminal process group ID" is switched to
the initial program.  If the broker that spawned it is
dispatched with a SIGKILL, flux-start receives a SIGTTOU.
The default action for SIGTTOU is STOP.

As discussed in #1831, flux-start can ignore SIGTTOU
to avoid being stopped right before exiting.