python: add --search option to make_binding.py
Problem: make_binding.py currently assumes all Flux API headers
are in the same directory, and this assumption doesn't hold for
generated headers like version.h, which may appear in a build
directory instead of source directory with other headers.

Add a new `--search` option to make_binding.py which can be used
to append to the script's search path, and internally build a list
of paths built from this path plus the "main" header path.