modules/kvs: Use zhashx_t in cache
Replace use of zhash_t with zhashx_t in cache.  By doing so,
we avoid duplicating the hash key/blobref that is already duplicated
in each cache entry.  The keys are also not duplicated when
cache entries are checked for expiration.

A global destructor for cache entries is also set, removing the need
to set a free function on every item in the hash.