libflux: abstract ready test for futures
Problem: several places in libflux/future.c test if a future
is ready or not ready by checking both f->result_valid *and*
f->fatal_errnum_valid. This requirement could too easily lead to a
future maintainer (hah) forgetting one of these checks, so abstract
this simple test into a convenience function and use it throughout
the code.

This change also cleans up `flux_future_is_ready()` to use the new
function. Though the function handily used `flux_future_wait_for (f, 0.)`
to test for readiness, in the end that amounted to the same check
implemented in the new `future_is_ready`, and use of that function
is more clear.