libflux/attr: move flags enum to broker
Problem: broker attr flags are defined in the public API,
but they are not used or needed.

Move them to the private broker attr implementation.

FLUX_ATTRFLAG_IMMUTABLE is needed internally by libflux/attr.c,
so just define that one internally with the same bit value
as the broker's definition.