libflux/attr: refactor internals
Clean up the implementation of the internal attribute cache.
The cache serves two purposes:
1) to store values forever that broker indicates are immutable
2) to store other values temporarily so that cache retains ownership
(freed the next time that attribute is looked up)

Rather than storing each value with a struct wrapper that
includes the flag indicating whether or not it is immutable
and putting all values in one zhash_t, create two zhashx_t's,
one for each purpose, and store string values directly in them.
The better size management of zhashx_t vs zhash_t should minimize
the impact of having two of them.  This allows internal code to
be simplified a bit.