libflux/module: drop lsmod codec
Problem: libflux/module defines functions for encoding
and decoding the lsmod response defined in RFC 5, and adds
a typedef to represent a list of module records.  The
abstraction is not really all that helpful with the other
API functions that are available and unnecessarily complicates
our API.

Drop these functions.

Update tests and users.