cmd/flux-module: list services in output
Add a "Services" column in flux module list output.
The column consists of any services returned from
the lsmod request that are not identical to the module
name, delimited by commas.

If a module registers different services on different ranks,
display each unique module + service as a separate line
of output.