libflux/module: add callback for dlopen/dlsym errors
Problem: flux_modname() logs to stderr when there is
a dlopen() or dlsym() error.

When searching a series of directories for a module to
load, it is helpful to not lose any dlopen() or dlsym()
errors that might provide clues as to why a particular
module cannot be loaded.

Since logging to stderr is not appropriate in a libflux
API function, extend flux_modname() and flux_modfind()
to allow an optional callback for dlopen error information,
if available.  The callback is simply informational,
and might be called via flux_modname() multiple times during
search, even if flux_modname() ultimately succeeds.

Update users: flux-module, broker, testsuite.