build: link cronodate test with libidset
Problem: the libutil/cronodate unit test needs to be
linked against libutil and libidset, but libidset needs
libutil/veb, so merely adding libidset to $(test_ldadd)
creates a circular dependency.

Make a special LDADD for this test program that specifies
the minimum set of objects, in the proper order to satisfy
dependencies.