testsuite: drop invalid valref watch returns nil tests
Problem: flux_kvs_watch_once[_dir]() seems to return nil,
not an error, when a malformed valref/diref object is
encountered.  We have tests that ensure this is the case.

The new implementation will fail when a malformed entry
is encountered, and this seems more correct.  Drop the
tests that check for a nil return.