python/makefile.am: remove unused and unrecognized variables
The `AM_LIBADD` variable is unrecognized by autotools.  It can be safely
removed since `PYTHON_LDFLAGS` are already included in the `common_libs`
variable, which is in-turn included in the `LIBADD` of all libraries'.

The _job variables and targets were vestiges from the reduction to a
single shared library in #1795.  They are unused and can be safely
removed as well.

Closes #1869