wreck: fix nokz with --output/error on i686
Problem: nokz jobs using --output, --error fail on i686 becuase
of some kind of sign error when converting FLUX_NODEID_ANY from
C to Lua to json and back out again. Therefore, the check in
the output.lua plugin fails to detect that the io service should
be registered on rank 0 of the job.

Explicitly use -1 instead of FLUX_NODEID_ANY to work around this
issue for 32bit systems.

Fixes #1895