testsuite: fix environment in valgrind test
Problem: flux module load in rc1 times out on slow
i686 system.

FLUX_LOCAL_CONNECTOR_RETRY_COUNT is set in the environment
of the test script in an attempt to compensate for slow
socket creation when running under valgrind, but the variable
is not inherited by the shell running rc1 script.

Add "export" to the variable setting.