travis-ci: fix github releases deployment
Problem: the travis-ci deploy condition was set to auto-deploy
for only the first Travis build, but now the first build is a
style check and doesn't have a deploy hook.

Instead of triggering the deploy off a build number, set a new
variable for the builder which is targeted for deployment.