testsuite: split submitbench out to test prog
Problem: the "submitbench" subcommand of flux job
does not really belong in a user-facing program.

Create a standalone, noinst program that contains
the ingest throughput test, then rename flux job
submitbench to flux job submit and simplify it so
that it only handles one job submission at a time.

In flux job submit, if no jobspec argument is present,
presume that it should be read from stdin, so
"flux job submit" is equivalent to  "flux job submit -".

Convert sharness tests to call either submitbench
or flux job submit as appropriate.

Cleanup:  now that travis-ci is starting the munge
daemon to support the default signing mech in
flux-security, there is no need to force all jobs
submitted in test to use --sign-type=none when
configured with --flux-security.