libjobspec: drop C++ jobspec library
Drop the C++ jobspec library, the 'validate-jobspec'
validator worker, sharness test, and the yaml-cpp
build dependency.

In addition, drop the configure.ac support for
configuring the C++ compiler, since there is no
other C++ code in the flux-core project.

The variant of the jobspec test that used this
validator is removed.

Updated build dependency table in README.md.