hwloc: use aggregator module for reduction and synchronization
Problem: The resource-hwloc module has no startup and reload
synchronization, so `flux module load resource-hwloc` as well as
`flux hwloc reload` may return before the module is ready and
toplogy is populated in the KVS.

Additionally, the `by_rank` directory hierarchy that the module
populates creates a lot of kvs keys and large transactions.

This change replaces the `by_rank` directory with a single JSON
object, aggregated for all identical ranks. The module uses the
aggregator to accomplish this, and thus also gains a synchronization
point. The module on rank 0 will always wait for the aggregate to
be "complete" before entering the reactor on startup, or responding
to RPC for a reload request.

All `resource-hwloc.reload` requests go to rank 0 now, and the
global reload is accomplished via a sequenced event that indicates
which ranks should actually reload topology information. All ranks,
however, re-perform the aggregation for any reload event.