testsuite: add jobmanager/sched-dummy
Add sched-dummy, an test schduler that doesn't
properly schedule anything, but demonstrates and
exercises the libschedutil code and the job manager
scheduler interface.

The scheduler is started with some number of imaginary
cores to allocate. Change with --cores option.

Each job is presumed to be requesting one core.
Parsing jobspec is beyond the capability of this scheduler.

When a core is available, a fake R is provided which
contains the string "1core".

There is a "module debug" flag (0x01) which causes allocations
to fail, so that this code can be exercised in the job-manager.