libflux/rpc: add FLUX_RPC_STREAMING flag
Problem: although logic exists to leak a matchtag if an RPC
future is destroyed without being fulfilled, this is insufficient
for streaming RPCs, which may still have responses in flight
even when the future is fulfilled.

Add a FLUX_RPC_STREAMING flag to designate an RPC as
expecting multiple responses, then leak the matchtag if it
has not been fulfilled with an error, which signifies
end-of-stream per RFC 6.