testsuite: valgrind: propagate workload script errors
Problem: The valgrind workoad script does not exit with error
if one of the workload scripts gets an error during execution
or fails to execute.

Emit an error on non-zero exit code from the workload scripts,
and ensure that the valgrind test itself fails on error from any
script, instead of exiting with success.